requirements.txt
pydantic
fastapi
uvicorn[standard]
python-dotenv
openai
python-multipart
pydantic
pika
ffmpeg-python
setuptools
openai-whisper
setuptools-rust
torch
supervisor